+++ src/components/LandingPage/MainPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './LandingPage.css';
+import './mainPage.css';
 
 const LandingPage = () => {
     return (
